Support non-junit-binding testing server and junit5 extension. (#25)
* Support non-junit-binding testing server and junit5 extension.

* Update CHANGES.md

* A simple workaround of Config object reuse and cache pollution issue.

* Move junit5 extension back to original mock server module.

* Move junit5 extension back to original mock server module.
