+++ CHANGES.md
@@ -6,5 +6,6 @@ Apollo Java 2.2.0
 
 ------------------
 [refactor(apollo-client): Optimize the exception message when failing to retrieve configuration information.](https://github.com/apolloconfig/apollo-java/pull/22)
+[Add JUnit5 extension support for apollo mock server.](https://github.com/apolloconfig/apollo-java/pull/25)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/2?closed=1)
