+++ apollo-mockserver/src/main/java/com/ctrip/framework/apollo/mockserver/MockApolloExtension.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2023 Apollo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package com.ctrip.framework.apollo.mockserver;
+
+import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
+import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
+
+import java.io.IOException;
+
+public class MockApolloExtension implements BeforeTestExecutionCallback, AfterTestExecutionCallback, ParameterResolver {
+
+    private ApolloTestingServer apollo = new ApolloTestingServer();
+
+    @Override
+    public void afterTestExecution(ExtensionContext context) {
+        apollo.close();
+    }
+
+    @Override
+    public void beforeTestExecution(ExtensionContext context) throws IOException {
+        apollo.start();
+    }
+
+    @Override
+    public boolean supportsParameter(
+        ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return parameterContext.getParameter().getType() == ApolloTestingServer.class
+            && extensionContext.getTestMethod().isPresent();
+    }
+
+    @Override
+    public Object resolveParameter(
+        ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return apollo;
+    }
+}
